# Conflicts:
#	CHANGES
#	cnchi/info.py

+++ cnchi/installation/auto_partition.py
@@ -114,34 +114,35 @@ def unmount_all(dest_dir):
         unmount(dest_dir)
 
 
-def remove_lvm():
+def remove_lvm(device):
     """ Remove all previous LVM volumes
     (it may have been left created due to a previous failed installation) """
-
     try:
-        lvolumes = check_output("lvs -o lv_name,vg_name --noheading").split("\n")
+        lvolumes = check_output("lvs -o lv_name,vg_name,devices --noheadings").split("\n")
         if len(lvolumes[0]) > 0:
             for lvolume in lvolumes:
                 if len(lvolume) > 0:
-                    (lvolume, vgroup) = lvolume.split()
-                    lvdev = "/dev/" + vgroup + "/" + lvolume
-                    subprocess.check_call(["wipefs", "-a", lvdev])
-                    subprocess.check_call(["lvremove", "-f", lvdev])
+                    (lvolume, vgroup, ldevice) = lvolume.split()
+                    if device in ldevice:
+                        lvdev = "/dev/" + vgroup + "/" + lvolume
+                        subprocess.check_call(["wipefs", "-a", lvdev])
+                        subprocess.check_call(["lvremove", "-f", lvdev])
 
-        vgnames = check_output("vgs -o vg_name --noheading").split("\n")
+        vgnames = check_output("vgs -o vg_name,devices --noheadings").split("\n")
         if len(vgnames[0]) > 0:
             for vgname in vgnames:
-                vgname = vgname.strip()
-                if len(vgname) > 0:
+                (vgname, vgdevice) = vgname.split()
+                if len(vgname) > 0 and device in vgdevice:
                     subprocess.check_call(["vgremove", "-f", vgname])
 
-        pvolumes = check_output("pvs -o pv_name --noheading").split("\n")
+        pvolumes = check_output("pvs -o pv_name --noheadings").split("\n")
         if len(pvolumes[0]) > 0:
             for pvolume in pvolumes:
                 pvolume = pvolume.strip(" ")
-                subprocess.check_call(["pvremove", "-f", pvolume])
+                if device in pvolume:
+                    subprocess.check_call(["pvremove", "-f", pvolume])
     except subprocess.CalledProcessError as err:
-        logging.warning("Can't delete existent LVM volumes, command %s failed: %s", err.cmd, err.output)
+        logging.warning("Can't delete existent LVM volumes in device %s, command %s failed: %s", device, err.cmd, err.output)
 
 
 def close_luks_devices():
